+++ bob/src/grinder.rs
@@ -17,6 +17,7 @@ pub struct Grinder {
     cluster: Arc<dyn Cluster + Send + Sync>,
     cleaner: Arc<Cleaner>,
     counter: Arc<BlobsCounter>,
+    node_config: NodeConfig,
     hw_counter: Arc<HWMetricsCollector>,
 }
 
@@ -45,6 +46,7 @@ impl Grinder {
             cluster: get_cluster(mapper, config, backend),
             cleaner,
             counter,
+            node_config: config.clone(),
             hw_counter,
         }
     }
@@ -55,6 +57,10 @@ impl Grinder {
 
     pub(crate) async fn run_backend(&self) -> Result<()> {
         self.backend.run_backend().await
+    }
+
+    pub(crate) fn node_config(&self) -> &NodeConfig {
+        &self.node_config
     }
 
     pub(crate) async fn put(