+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Bob versions changelog
 
 
 #### Changed
-
+- Disable AIO by default (#335)
 
 
 #### Fixed