+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Bob versions changelog
 
 
 #### Fixed
-
+- Hardware metrics names for RAM and disk space (#358)
 
 
 #### Updated