+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Bob versions changelog
 ## [Unreleased]
 #### Added
 - Add brt mode to reverse byte order ([#286](https://github.com/qoollo/bob/pull/286))
-
+- Add alien blobs sync API ([#334](https://github.com/qoollo/bob/pull/346))
 
 
 #### Changed