+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Bob versions changelog
 
 
 #### Updated
+- Upgrade pearl to v0.8.0